Move extensions under SAGESharp.SLB to SAGESharp.Utils
