+++ SAGESharp/SAGESharp.csproj
@@ -125,7 +125,6 @@
     <Compile Include="OSI\OSIFile.cs" />
     <Compile Include="OSI\Instruction.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="SLB\ASCIIExtensions.cs" />
     <Compile Include="SLB\Cinematic\Camera\Camera.cs" />
     <Compile Include="SLB\Cinematic\Camera\Frame.cs" />
     <Compile Include="SLB\Common.cs" />
@@ -141,9 +140,10 @@
     <Compile Include="SLB\Level\Conversation\Info.cs" />
     <Compile Include="SLB\Level\Conversation\InfoBinaryWriter.cs" />
     <Compile Include="SLB\Level\Conversation\LineSide.cs" />
-    <Compile Include="SLB\NumericExtensions.cs" />
-    <Compile Include="SLB\StreamExtensions.cs" />
     <Compile Include="SLB\StringBinaryWriter.cs" />
+    <Compile Include="Utils\ASCIIExtensions.cs" />
+    <Compile Include="Utils\NumericExtensions.cs" />
+    <Compile Include="Utils\StreamExtensions.cs" />
     <Compile Include="Validations.cs" />
     <Compile Include="XFile.cs" />
     <Compile Include="XUtils.cs" />
