+++ SAGESharpTests/SAGESharpTests.csproj
@@ -88,7 +88,6 @@
     <Compile Include="IO\PrimitiveBinarySerializerTests.cs" />
     <Compile Include="IO\StringBinarySerializerTests.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="SLB\ASCIIExtensionsTests.cs" />
     <Compile Include="SLB\Cinematic\Camera\CameraEqualityTests.cs" />
     <Compile Include="SLB\Cinematic\Camera\FrameEqualityTests.cs" />
     <Compile Include="SLB\FooterEntryTests.cs" />
@@ -106,17 +105,18 @@
     <Compile Include="SLB\Level\Conversation\InfoEqualityTests.cs" />
     <Compile Include="SLB\Level\Conversation\SerializationTests.cs" />
     <Compile Include="SLB\Level\Conversation\TestData.cs" />
-    <Compile Include="SLB\NumericExtensionsTests.cs" />
     <Compile Include="SLB\Point3DEqualityTests.cs" />
     <Compile Include="SLB\SLBFooterWriterTests.cs" />
     <Compile Include="SLB\StringBinaryWriterTests.cs" />
-    <Compile Include="SLB\StreamExtensionsTests.cs" />
     <Compile Include="Testing\AbstractTestCaseData.cs" />
     <Compile Include="Testing\Matchers.cs" />
     <Compile Include="Testing\ParameterGroup.cs" />
     <Compile Include="Testing\AbstractEqualityByRefTests.cs" />
     <Compile Include="Testing\AbstractEqualityByValTests.cs" />
     <Compile Include="Testing\SerializationTestCaseData.cs" />
+    <Compile Include="Utils\ASCIIExtensionsTests.cs" />
+    <Compile Include="Utils\NumericExtensionsTests.cs" />
+    <Compile Include="Utils\StreamExtensionsTests.cs" />
     <Compile Include="ValidationsTests.cs" />
   </ItemGroup>
   <ItemGroup>
